renamed all pictures to images (#22)
