+++ src/fool_ai_detector/service/naive_baseline_processor_image.py
